requirements.txt
pysam
biopython
mosdepth
samtools
sniffles
bcftools
